Wk2 Minimum Requirements Met
Met the Wk2 Minimum Requirements.

Although you can’t select the status on the “Create New Project” —
there is a hidden input with an id of Status, and by default, it will
go into the Discovery phase automatically. There’s accordion tabs for
the menu, and on the home page, for registration, I’ve included the
tooltips.